Late RTC sync fix
The issue was present with my first Pi connected to Internet via WiFi. As the connection is estabilished later in boot time, so is the RTC synced later.
When the clock was adjusted, the pistackomd's real time process hogged CPU for about 20 minutes to catch up with 50 years worth of sw PWM with no sleep.